1. move load and classify funcs to BoaMLIntrinsics. 2. use job id and model var name as model identifier.
