+++ src/java/boa/compiler/SymbolTable.java
@@ -556,6 +556,7 @@ public class SymbolTable {
 			boa.functions.BoaAstIntrinsics.class,
 			boa.functions.BoaGraphIntrinsics.class,
 			boa.functions.BoaIntrinsics.class,
+			boa.functions.BoaMLIntrinsics.class,
 			boa.functions.BoaMetricIntrinsics.class,
 			boa.functions.BoaNormalFormIntrinsics.class,
 			boa.functions.BoaModifierIntrinsics.class,
