+++ src/java/boa/functions/BoaIntrinsics.java
@@ -16,11 +16,6 @@
  */
 package boa.functions;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.ObjectInputStream;
-import java.lang.reflect.Array;
-import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -33,27 +28,11 @@ import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.apache.commons.lang.math.NumberUtils;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FSDataInputStream;
-import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.MapFile;
-import org.apache.hadoop.mapreduce.Mapper.Context;
-
-import boa.datagen.DefaultProperties;
 import boa.types.Code.CodeRepository;
 import boa.types.Code.Revision;
 import boa.types.Diff.ChangedFile;
 import boa.types.Shared.ChangeKind;
 import boa.types.Toplevel.Project;
-import boa.types.ml.BoaLinearRegression;
-import boa.types.ml.BoaModel;
-import weka.classifiers.Classifier;
-import weka.core.Attribute;
-import weka.core.DenseInstance;
-import weka.core.Instance;
-import weka.core.Instances;
 
 /**
  * Boa domain-specific functions.
@@ -101,85 +80,7 @@ public class BoaIntrinsics {
 		}
 		return -1;
 	}
-	
-	/**
-	 * Given the model URL, deserialize the model and return Model type
-	 *
-	 * @param Take URL for the model
-	 * @return Model type after deserializing
-	 */
-	@FunctionSpec(name = "load", returnType = "LinearRegression", formalParameters = {"string"})
-	public static BoaModel load(final String jobId, final Object o) throws Exception {
-		Object unserializedObject = null;
-		FSDataInputStream in = null;
-		ObjectInputStream dataIn = null;
-		ByteArrayOutputStream bo = null;
-		try {
-			final Configuration conf = BoaAstIntrinsics.context.getConfiguration();
-			final FileSystem fs;
-			final Path p;
-			String output = DefaultProperties.localOutput != null ? new Path(DefaultProperties.localOutput).toString() + "/../"
-					: conf.get("fs.default.name", "hdfs://boa-njt/");
-			p = new Path(output, new Path("model/job_" + jobId + ".model"));
-			fs = FileSystem.get(conf);
-			in = fs.open(p);
-			
-			final byte[] b = new byte[(int)fs.getLength(p) + 1];
-			int c = 0;
-			bo = new ByteArrayOutputStream();
-			while((c = in.read(b)) != -1){
-				bo.write(b, 0, c);
-			}
-			ByteArrayInputStream bin = new ByteArrayInputStream(bo.toByteArray());
-			dataIn = new ObjectInputStream(bin);
-			unserializedObject = dataIn.readObject();
-		}
-		catch(Exception e){
-			e.printStackTrace();
-		}
-		finally {
-			try {
-				if (in != null) in.close();
-				if (dataIn != null) dataIn.close();
-				if (bo != null) bo.close();
-			} catch (final Exception e) { e.printStackTrace(); }
-		}
 
-		Classifier clr = (Classifier)unserializedObject;
-		BoaModel m = null;
-		if(clr.toString().contains("Linear Regression")){
-			m = new BoaLinearRegression(clr, o);
-		}
-		return m;
-	}
-	
-	
-	@FunctionSpec(name = "classify", returnType = "string", formalParameters = { "Model","array of int"})
-	public static String classify(final BoaModel model, final long[] vector) throws Exception {
-		int NumOfAttributes = vector.length + 1;
-		ArrayList<Attribute> fvAttributes = new ArrayList<Attribute>();
-	
-		for(int i=0; i < NumOfAttributes - 1; i++) {
-			fvAttributes.add(new Attribute("Attribute" + i));
-		}
-
-		Instances testingSet = new Instances("Classifier", fvAttributes, 1);
-		testingSet.setClassIndex(testingSet.numAttributes() - 1);
-		Instance instance = new DenseInstance(NumOfAttributes - 1);
-		for(int i=0; i<NumOfAttributes-1; i++) {
-			instance.setValue((Attribute)fvAttributes.get(i), vector[i]);
-		}
-		testingSet.add(instance);
-
-		Classifier classifier = (Classifier) model.getClassifier();
-		double predval = classifier.classifyInstance(testingSet.instance(0));
-		
-		if(testingSet.classAttribute().isNominal())
-			return testingSet.classAttribute().value((int) predval);
-		else
-			return predval + "";
-	}
-	
 	@FunctionSpec(name = "getrevisionscount", returnType = "int", formalParameters = { "CodeRepository" })
 	public static int getRevisionsCount(CodeRepository cr) {
 		return Math.max(cr.getRevisionKeysCount(), cr.getRevisionsCount());
