+++ test/ml-programs/linear-regression/classesatrevision/classes_at_revision_test.boa
@@ -7,7 +7,7 @@ totalloc:=0;
 collect : output collection[string][int] of string;
 
 model: Model of LinearRegression of {int,int};
-model = load(path/to/model);
+model = load(job_id: int, model_var: string);
 
 lloc := function(node: ASTRoot) : int {
     count := 0;
